CrowdCounting-on-VisDrone2020-main/requirements.txt
torch==1.10.1+cu113
numpy~=1.21.5
torchvision==0.11.2+cu113
pandas~=1.2.5
h5py~=3.6.0
scikit-learn~=1.0.1
Pillow~=8.2.0
easydict~=1.9
opencv-python~=4.5.4.60
matplotlib~=3.4.3
scipy~=1.7.3
tqdm~=4.60.0
tensorboardX~=2.2
tabulate~=0.8.9
hdbscan~=0.8.27
torchaudio==0.10.1+cu113
wheel~=0.36.2
